Deletion of settings page
We made many attempts to integrate changing the attributes of other forms from the settings form but Java provides very limited funcitonality for this. As such, we have deleted it for the sake of the rest of our program. The use of high contrast with our colours and text still makes it readable for people with colourblindness, the colours just no longer change to represent what we see to them. Also, we will be adding clear and large text everywhere so there is no need to magnify it.